[Task 05] Organizando projeto
